requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
# LangChain and LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph-checkpoint
langgraph-sdk>=0.1.0
google-generativeai
openai
anthropic
chromadb
tiktoken
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
redis
celery
pytest
pytest-asyncio
httpx
websockets
streamlit
PyPDF2
aiosqlite
# Production improvements
slowapi
tenacity
python-json-logger
prometheus-client
# Web search
ddgs
# Web scraping
httpx
beautifulsoup4
lxml
# Document processing
python-docx
